Implement code sharing
This commit contains all the changes to store the code and then generate a URL
that can be shared, the current implementation can store in local for development
or in dropbox which is a free alternative.

In future will be implemented a new feature to store and download to S3 bucket.