Add Image controller and refactor file service
FileService now return a SystemFile instead an int[]
this allow to use the download directly for code and images.

This commit also include the logic to generate the SEO with the
screenshot of the code when loading, this will allow when share
it will show an screen capture of the current code.